+++ frameworks/Rust/axum/src/main_mongo.rs
@@ -53,7 +53,7 @@ async fn queries(DatabaseConnection(mut db): DatabaseConnection, Query(params): 
 
     let mut rng = SmallRng::from_entropy();
 
-    let mut results = Vec::with_capacity(q as usize);
+    let mut results = Vec::with_capacity(q);
 
     for _ in 0..q {
         let query_id = random_number(&mut rng);
